Replace CreateNetworksUtil & NetworkFactory by Networks class
