+++ diagram-test/src/main/java/com/powsybl/diagram/test/Networks.java
@@ -6,7 +6,9 @@
  */
 package com.powsybl.diagram.test;
 
+import com.powsybl.commons.extensions.Extendable;
 import com.powsybl.iidm.network.*;
+import com.powsybl.iidm.network.extensions.*;
 import org.joda.time.DateTime;
 
 import java.util.Optional;
@@ -919,5 +921,1005 @@ public final class Networks {
                 .setB(0)
                 .add();
         return network;
+    }
+
+    public static Network createTestCase11Network() {
+        Network network = Network.create("testCase11", "test");
+
+        Substation substation = createSubstation(network, "subst", "subst", Country.FR);
+
+        // first voltage level
+        //
+        VoltageLevel vl1 = createVoltageLevel(substation, "vl1", "vl1", TopologyKind.NODE_BREAKER, 380, 50);
+
+        createBusBarSection(vl1, "bbs1", "bbs1", 0, 1, 1);
+        createBusBarSection(vl1, "bbs2", "bbs2", 1, 1, 2);
+        createBusBarSection(vl1, "bbs3", "bbs3", 2, 2, 1);
+        createBusBarSection(vl1, "bbs4", "bbs4", 3, 2, 2);
+
+        createSwitch(vl1, "dsect11", "dsect11", SwitchKind.DISCONNECTOR, false, false, true, 0, 14);
+        createSwitch(vl1, "dtrct11", "dtrct11", SwitchKind.BREAKER, true, false, true, 14, 15);
+        createSwitch(vl1, "dsect12", "dsect12", SwitchKind.DISCONNECTOR, false, false, true, 15, 1);
+
+        createSwitch(vl1, "dsect21", "dsect21", SwitchKind.DISCONNECTOR, false, false, true, 2, 16);
+        createSwitch(vl1, "dtrct21", "dtrct21", SwitchKind.BREAKER, true, false, true, 16, 17);
+        createSwitch(vl1, "dsect22", "dsect22", SwitchKind.DISCONNECTOR, false, false, true, 17, 3);
+
+        createLoad(vl1, "load1", "load1", "load1", 0, ConnectablePosition.Direction.TOP, 4, 10, 10);
+        createSwitch(vl1, "dload1", "dload1", SwitchKind.DISCONNECTOR, false, false, true, 0, 5);
+        createSwitch(vl1, "bload1", "bload1", SwitchKind.BREAKER, true, false, true, 4, 5);
+
+        createGenerator(vl1, "gen1", "gen1", "gen1", 2, ConnectablePosition.Direction.BOTTOM, 6, 0, 20, false, 10, 10);
+        createSwitch(vl1, "dgen1", "dgen1", SwitchKind.DISCONNECTOR, false, false, true, 2, 7);
+        createSwitch(vl1, "bgen1", "bgen1", SwitchKind.BREAKER, true, false, true, 6, 7);
+
+        createLoad(vl1, "load2", "load2", "load2", 8, ConnectablePosition.Direction.TOP, 8, 10, 10);
+        createSwitch(vl1, "dload2", "dload2", SwitchKind.DISCONNECTOR, false, false, true, 1, 9);
+        createSwitch(vl1, "bload2", "bload2", SwitchKind.BREAKER, true, false, true, 8, 9);
+
+        createGenerator(vl1, "gen2", "gen2", "gen2", 12, ConnectablePosition.Direction.BOTTOM, 10, 0, 20, false, 10, 10);
+        createSwitch(vl1, "dgen2", "dgen2", SwitchKind.DISCONNECTOR, false, false, true, 3, 11);
+        createSwitch(vl1, "bgen2", "bgen2", SwitchKind.BREAKER, true, false, true, 10, 11);
+
+        // second voltage level
+        //
+        VoltageLevel vl2 = createVoltageLevel(substation, "vl2", "vl2", TopologyKind.NODE_BREAKER, 225, 30);
+
+        createBusBarSection(vl2, "bbs5", "bbs5", 0, 1, 1);
+        createBusBarSection(vl2, "bbs6", "bbs6", 1, 2, 1);
+
+        createSwitch(vl2, "dscpl1", "dscpl1", SwitchKind.DISCONNECTOR, false, false, true, 0, 6);
+        createSwitch(vl2, "ddcpl1", "ddcpl1", SwitchKind.BREAKER, true, false, true, 6, 7);
+        createSwitch(vl2, "dscpl2", "dscpl2", SwitchKind.DISCONNECTOR, false, false, true, 7, 1);
+
+        createLoad(vl2, "load3", "load3", "load3", 0, ConnectablePosition.Direction.TOP, 2, 10, 10);
+        createSwitch(vl2, "dload3", "dload3", SwitchKind.DISCONNECTOR, false, false, true, 0, 3);
+        createSwitch(vl2, "bload3", "bload3", SwitchKind.BREAKER, true, false, true, 2, 3);
+
+        createGenerator(vl2, "gen4", "gen4", "gen4", 2, ConnectablePosition.Direction.BOTTOM, 4, 0, 20, false, 10, 10);
+        createSwitch(vl2, "dgen4", "dgen4", SwitchKind.DISCONNECTOR, false, false, true, 1, 5);
+        createSwitch(vl2, "bgen4", "bgen4", SwitchKind.BREAKER, true, false, true, 4, 5);
+
+        // third voltage level
+        //
+        VoltageLevel vl3 = createVoltageLevel(substation, "vl3", "vl3", TopologyKind.NODE_BREAKER, 50, 20);
+
+        createBusBarSection(vl3, "bbs7", "bbs7", 0, 1, 1);
+
+        createLoad(vl3, "load4", "load4", "load4", 0, ConnectablePosition.Direction.TOP, 1, 10, 10);
+        createSwitch(vl3, "dload4", "dload4", SwitchKind.DISCONNECTOR, false, false, true, 0, 2);
+        createSwitch(vl3, "bload4", "bload4", SwitchKind.BREAKER, true, false, true, 2, 1);
+
+        // two windings transformers between voltage levels
+        //
+        createSwitch(vl1, "dtrf11", "dtrf11", SwitchKind.DISCONNECTOR, false, false, true, 0, 18);
+        createSwitch(vl1, "btrf11", "btrf11", SwitchKind.BREAKER, true, false, true, 18, 19);
+        createSwitch(vl2, "dtrf21", "dtrf21", SwitchKind.DISCONNECTOR, false, false, true, 0, 8);
+        createSwitch(vl2, "btrf21", "btrf21", SwitchKind.BREAKER, true, false, true, 8, 9);
+        createTwoWindingsTransformer(substation, "trf1", "trf1", 2.0, 14.745, 0.0, 3.2E-5, 400.0, 225.0,
+                19, 9, vl1.getId(), vl2.getId(),
+                "trf1", 1, ConnectablePosition.Direction.TOP,
+                "trf1", 1, ConnectablePosition.Direction.TOP);
+
+        createSwitch(vl1, "dtrf12", "dtrf12", SwitchKind.DISCONNECTOR, false, false, true, 1, 20);
+        createSwitch(vl1, "btrf12", "btrf12", SwitchKind.BREAKER, true, false, true, 20, 21);
+        createSwitch(vl2, "dtrf22", "dtrf22", SwitchKind.DISCONNECTOR, false, false, true, 1, 10);
+        createSwitch(vl2, "btrf22", "btrf22", SwitchKind.BREAKER, true, false, true, 10, 11);
+        createTwoWindingsTransformer(substation, "trf2", "trf2", 2.0, 14.745, 0.0, 3.2E-5, 400.0, 225.0,
+                21, 11, vl1.getId(), vl2.getId(),
+                "trf2", 11, ConnectablePosition.Direction.TOP,
+                "trf2", 7, ConnectablePosition.Direction.BOTTOM);
+
+        createSwitch(vl1, "dtrf13", "dtrf13", SwitchKind.DISCONNECTOR, false, false, true, 2, 22);
+        createSwitch(vl1, "btrf13", "btrf13", SwitchKind.BREAKER, true, false, true, 22, 23);
+        createSwitch(vl2, "dtrf23", "dtrf23", SwitchKind.DISCONNECTOR, false, false, true, 1, 12);
+        createSwitch(vl2, "btrf23", "btrf23", SwitchKind.BREAKER, true, false, true, 12, 13);
+        createTwoWindingsTransformer(substation, "trf3", "trf3", 2.0, 14.745, 0.0, 3.2E-5, 400.0, 225.0,
+                23, 13, vl1.getId(), vl2.getId(),
+                "trf3", 3, ConnectablePosition.Direction.BOTTOM,
+                "trf3", 8, ConnectablePosition.Direction.BOTTOM);
+
+        createSwitch(vl1, "dtrf14", "dtrf14", SwitchKind.DISCONNECTOR, false, false, true, 3, 24);
+        createSwitch(vl1, "btrf14", "btrf14", SwitchKind.BREAKER, true, false, true, 24, 25);
+        createSwitch(vl2, "dtrf24", "dtrf24", SwitchKind.DISCONNECTOR, false, false, true, 0, 14);
+        createSwitch(vl2, "btrf24", "btrf24", SwitchKind.BREAKER, true, false, true, 14, 15);
+        createTwoWindingsTransformer(substation, "trf4", "trf4", 2.0, 14.745, 0.0, 3.2E-5, 400.0, 225.0,
+                25, 15, vl1.getId(), vl2.getId(),
+                "trf4", 10, ConnectablePosition.Direction.BOTTOM,
+                "trf4", 3, ConnectablePosition.Direction.TOP);
+
+        createSwitch(vl1, "dtrf15", "dtrf15", SwitchKind.DISCONNECTOR, false, false, true, 0, 26);
+        createSwitch(vl1, "btrf15", "btrf15", SwitchKind.BREAKER, true, false, true, 26, 27);
+        createSwitch(vl3, "dtrf25", "dtrf25", SwitchKind.DISCONNECTOR, false, false, true, 0, 3);
+        createSwitch(vl3, "btrf25", "btrf25", SwitchKind.BREAKER, true, false, true, 3, 4);
+        createTwoWindingsTransformer(substation, "trf5", "trf5", 2.0, 14.745, 0.0, 3.2E-5, 400.0, 225.0,
+                27, 4, vl1.getId(), vl3.getId(),
+                "trf5", 4, ConnectablePosition.Direction.TOP,
+                "trf5", 1, ConnectablePosition.Direction.BOTTOM);
+
+        // three windings transformers between voltage levels
+        //
+        createSwitch(vl1, "dtrf16", "dtrf16", SwitchKind.DISCONNECTOR, false, false, true, 0, 28);
+        createSwitch(vl1, "btrf16", "btrf16", SwitchKind.BREAKER, true, false, true, 28, 29);
+        createSwitch(vl2, "dtrf26", "dtrf26", SwitchKind.DISCONNECTOR, false, false, true, 1, 16);
+        createSwitch(vl2, "btrf26", "btrf26", SwitchKind.BREAKER, true, false, true, 16, 17);
+        createSwitch(vl3, "dtrf36", "dtrf36", SwitchKind.DISCONNECTOR, false, false, true, 0, 5);
+        createSwitch(vl3, "btrf36", "btrf36", SwitchKind.BREAKER, true, false, true, 5, 6);
+
+        createThreeWindingsTransformer(substation, "trf6", "trf6", vl1.getId(), vl2.getId(), vl3.getId(),
+                0.5, 0.5, 0.5, 1., 1., 1., 0.1, 0.1,
+                400., 225., 50.,
+                29, 17, 6,
+                "trf61", 5, ConnectablePosition.Direction.TOP,
+                "trf62", 5, ConnectablePosition.Direction.TOP,
+                "trf63", 2, ConnectablePosition.Direction.TOP);
+
+        createSwitch(vl1, "dtrf17", "dtrf17", SwitchKind.DISCONNECTOR, false, false, true, 2, 30);
+        createSwitch(vl1, "btrf17", "btrf17", SwitchKind.BREAKER, true, false, true, 30, 31);
+        createSwitch(vl2, "dtrf27", "dtrf27", SwitchKind.DISCONNECTOR, false, false, true, 0, 18);
+        createSwitch(vl2, "btrf27", "btrf27", SwitchKind.BREAKER, true, false, true, 18, 19);
+        createSwitch(vl3, "dtrf37", "dtrf37", SwitchKind.DISCONNECTOR, false, false, true, 0, 7);
+        createSwitch(vl3, "btrf37", "btrf37", SwitchKind.BREAKER, true, false, true, 7, 8);
+
+        createThreeWindingsTransformer(substation, "trf7", "trf7", vl1.getId(), vl2.getId(), vl3.getId(),
+                0.5, 0.5, 0.5, 1., 1., 1., 0.1, 0.1,
+                400., 225., 50.,
+                31, 19, 8,
+                "trf71", 6, ConnectablePosition.Direction.BOTTOM,
+                "trf72", 4, ConnectablePosition.Direction.TOP,
+                "trf73", 3, ConnectablePosition.Direction.BOTTOM);
+
+        createSwitch(vl1, "dtrf18", "dtrf18", SwitchKind.DISCONNECTOR, false, false, true, 1, 32);
+        createSwitch(vl1, "btrf18", "btrf18", SwitchKind.BREAKER, true, false, true, 32, 33);
+        createSwitch(vl2, "dtrf28", "dtrf28", SwitchKind.DISCONNECTOR, false, false, true, 1, 20);
+        createSwitch(vl2, "btrf28", "btrf28", SwitchKind.BREAKER, true, false, true, 20, 21);
+        createSwitch(vl3, "dtrf38", "dtrf38", SwitchKind.DISCONNECTOR, false, false, true, 0, 9);
+        createSwitch(vl3, "btrf38", "btrf38", SwitchKind.BREAKER, true, false, true, 9, 10);
+
+        createThreeWindingsTransformer(substation, "trf8", "trf8", vl1.getId(), vl2.getId(), vl3.getId(),
+                0.5, 0.5, 0.5, 1., 1., 1., 0.1, 0.1,
+                400., 225., 50.,
+                33, 21, 10,
+                "trf81", 9, ConnectablePosition.Direction.TOP,
+                "trf82", 6, ConnectablePosition.Direction.BOTTOM,
+                "trf83", 4, ConnectablePosition.Direction.TOP);
+
+        // Creation of another substation, another voltageLevel and a line between the two substations
+        //
+        Substation substation2 = createSubstation(network, "subst2", "subst2", Country.FR);
+        VoltageLevel vlSubst2 = createVoltageLevel(substation2, "vlSubst2", "vlSubst2", TopologyKind.NODE_BREAKER, 380, 50);
+
+        createBusBarSection(vlSubst2, "bbs1_2", "bbs1_2", 0, 1, 1);
+
+        createSwitch(vl1, "dline11_2", "dline11_2", SwitchKind.DISCONNECTOR, false, false, true, 0, 34);
+        createSwitch(vl1, "bline11_2", "bline11_2", SwitchKind.BREAKER, true, false, true, 34, 35);
+        createSwitch(vlSubst2, "dline21_2", "dline21_2", SwitchKind.DISCONNECTOR, false, false, true, 0, 1);
+        createSwitch(vlSubst2, "bline21_2", "bline21_2", SwitchKind.BREAKER, true, false, true, 1, 2);
+        createLine(network, "line1", "line1", 2.0, 14.745, 1.0, 1.0, 1.0, 1.0,
+                35, 2, vl1.getId(), vlSubst2.getId(),
+                "line1", 7, ConnectablePosition.Direction.TOP,
+                "line1", 1, ConnectablePosition.Direction.TOP);
+
+        return network;
+    }
+
+    public static Network createNodeBreakerNetworkWithBranchStatus(String id, String sourceFormat) {
+        Network network = createNodeBreakerNetworkWithInternalBranches(id, sourceFormat);
+
+        createStatusExtensions(network);
+
+        return network;
+    }
+
+    public static Network createBusBreakerNetworkWithBranchStatus(String id, String sourceFormat) {
+        Network network = createBusBreakerNetworkWithInternalBranches(id, sourceFormat);
+
+        createStatusExtensions(network);
+
+        return network;
+    }
+
+    private static void createStatusExtensions(Network network) {
+        network.getLine("L11").newExtension(BranchStatusAdder.class).withStatus(BranchStatus.Status.PLANNED_OUTAGE).add();
+        network.getLine("L12").newExtension(BranchStatusAdder.class).withStatus(BranchStatus.Status.FORCED_OUTAGE).add();
+
+        network.getTwoWindingsTransformer("T11").newExtension(BranchStatusAdder.class).withStatus(BranchStatus.Status.PLANNED_OUTAGE).add();
+        network.getTwoWindingsTransformer("T12").newExtension(BranchStatusAdder.class).withStatus(BranchStatus.Status.FORCED_OUTAGE).add();
+
+        network.getThreeWindingsTransformer("T3_12").newExtension(BranchStatusAdder.class).withStatus(BranchStatus.Status.FORCED_OUTAGE).add();
+    }
+
+    public static Network createNodeBreakerNetworkWithInternalBranches(String id, String sourceFormat) {
+        Network network = Network.create(id, sourceFormat);
+        Substation substation = createSubstation(network, "S1", "S1", Country.FR);
+
+        VoltageLevel vl1 = createVoltageLevel(substation, "VL1", "VL1", TopologyKind.NODE_BREAKER, 400, 24);
+
+        createBusBarSection(vl1, "BBS11", "BBS11", 10, 1, 1);
+        createBusBarSection(vl1, "BBS12", "BBS12", 20, 2, 1);
+
+        createSwitch(vl1, "D10", "D10", SwitchKind.DISCONNECTOR, false, false, false, 10, 120);
+        createSwitch(vl1, "BR1", "BR1", SwitchKind.BREAKER, false, false, false, 120, 121);
+        createSwitch(vl1, "D20", "D20", SwitchKind.DISCONNECTOR, false, false, false, 121, 20);
+
+        createLoad(vl1, "L1", "L1", "L1", null, ConnectablePosition.Direction.TOP, 12, 1, 1);
+
+        createSwitch(vl1, "D11", "D11", SwitchKind.DISCONNECTOR, false, false, false, 10, 11);
+        createSwitch(vl1, "BR12", "BR12", SwitchKind.BREAKER, false, false, false, 11, 12);
+
+        createGenerator(vl1, "G", "G", "G", null, ConnectablePosition.Direction.TOP, 22, 50, 100, false, 100, 400);
+
+        createSwitch(vl1, "D21", "D21", SwitchKind.DISCONNECTOR, false, false, false, 20, 21);
+        createSwitch(vl1, "BR22", "BR22", SwitchKind.BREAKER, false, false, false, 21, 22);
+
+        createSwitch(vl1, "D13", "D13", SwitchKind.DISCONNECTOR, false, false, false, 10, 13);
+        createSwitch(vl1, "BR14", "BR14", SwitchKind.BREAKER, false, false, false, 13, 14);
+
+        createSwitch(vl1, "D23", "D23", SwitchKind.DISCONNECTOR, false, false, false, 20, 23);
+        createSwitch(vl1, "BR24", "BR24", SwitchKind.BREAKER, false, false, false, 23, 24);
+
+        createLine(network, "L11", "L11", 1.0, 1.0, 1.0, 0.0, 0.0, 0.0,
+                14, 24, vl1.getId(), vl1.getId(),
+                "L11", null, ConnectablePosition.Direction.TOP,
+                "L11", null, ConnectablePosition.Direction.TOP);
+
+        VoltageLevel vl2 = createVoltageLevel(substation, "VL2", "VL2", TopologyKind.NODE_BREAKER, 400, 9);
+
+        createBusBarSection(vl2, "BBS2", "BBS2", 30, 1, 1);
+
+        createLoad(vl2, "L2", "L2", "L2", null, ConnectablePosition.Direction.TOP, 32, 1, 1);
+
+        createSwitch(vl2, "D31", "D31", SwitchKind.DISCONNECTOR, false, false, false, 30, 31);
+        createSwitch(vl2, "BR32", "BR32", SwitchKind.BREAKER, false, false, false, 31, 32);
+
+        createSwitch(vl1, "D15", "D15", SwitchKind.DISCONNECTOR, false, false, false, 10, 15);
+        createSwitch(vl1, "BR16", "BR16", SwitchKind.BREAKER, false, false, false, 15, 16);
+
+        createSwitch(vl2, "D33", "D33", SwitchKind.DISCONNECTOR, false, false, false, 30, 33);
+        createSwitch(vl2, "BR34", "BR34", SwitchKind.BREAKER, false, false, false, 33, 34);
+
+        createLine(network, "L12", "L12", 1.0, 1.0, 1.0, 0.0, 0.0, 0.0,
+                16, 34, vl1.getId(), vl2.getId(),
+                "L12", null, ConnectablePosition.Direction.TOP,
+                "L12", null, ConnectablePosition.Direction.TOP);
+
+        createSwitch(vl1, "D17", "D17", SwitchKind.DISCONNECTOR, false, false, false, 10, 17);
+        createSwitch(vl1, "BR18", "BR18", SwitchKind.BREAKER, false, false, false, 17, 18);
+
+        createSwitch(vl1, "D25", "D25", SwitchKind.DISCONNECTOR, false, false, false, 20, 25);
+        createSwitch(vl1, "BR26", "BR26", SwitchKind.BREAKER, false, false, false, 25, 26);
+
+        createTwoWindingsTransformer(substation, "T11", "T11", 250, 100, 52, 12, 65, 90,
+                18, 26, vl1.getId(), vl1.getId(),
+                "T11", null, ConnectablePosition.Direction.TOP,
+                "T11", null, ConnectablePosition.Direction.TOP);
+
+        TwoWindingsTransformer twoWindingsTransformer = network.getTwoWindingsTransformer("T11");
+        twoWindingsTransformer.getTerminal(TwoWindingsTransformer.Side.ONE).setP(375);
+        twoWindingsTransformer.getTerminal(TwoWindingsTransformer.Side.TWO).setP(375);
+        twoWindingsTransformer.getTerminal(TwoWindingsTransformer.Side.ONE).setQ(48);
+        twoWindingsTransformer.getTerminal(TwoWindingsTransformer.Side.TWO).setQ(48);
+
+        createSwitch(vl1, "D27", "D27", SwitchKind.DISCONNECTOR, false, false, false, 20, 27);
+        createSwitch(vl1, "BR28", "BR28", SwitchKind.BREAKER, false, false, false, 27, 28);
+
+        createSwitch(vl2, "D35", "D35", SwitchKind.DISCONNECTOR, false, false, false, 30, 35);
+        createSwitch(vl2, "BR36", "BR36", SwitchKind.BREAKER, false, false, false, 35, 36);
+
+        createTwoWindingsTransformer(substation, "T12", "T12", 250, 100, 52, 12, 65, 90,
+                28, 36, vl1.getId(), vl2.getId(),
+                "T12", null, ConnectablePosition.Direction.TOP,
+                "T12", null, ConnectablePosition.Direction.TOP);
+
+        twoWindingsTransformer = network.getTwoWindingsTransformer("T12");
+        twoWindingsTransformer.getTerminal(TwoWindingsTransformer.Side.ONE).setP(375);
+        twoWindingsTransformer.getTerminal(TwoWindingsTransformer.Side.TWO).setP(375);
+        twoWindingsTransformer.getTerminal(TwoWindingsTransformer.Side.ONE).setQ(48);
+        twoWindingsTransformer.getTerminal(TwoWindingsTransformer.Side.TWO).setQ(48);
+
+        createSwitch(vl1, "D19", "D19", SwitchKind.DISCONNECTOR, false, false, false, 10, 19);
+        createSwitch(vl1, "BR20", "BR20", SwitchKind.BREAKER, false, false, false, 19, 190);
+
+        createSwitch(vl1, "D29", "D29", SwitchKind.DISCONNECTOR, false, false, false, 20, 29);
+        createSwitch(vl1, "BR30", "BR30", SwitchKind.BREAKER, false, false, false, 29, 290);
+
+        createSwitch(vl2, "D37", "D37", SwitchKind.DISCONNECTOR, false, false, false, 30, 37);
+        createSwitch(vl2, "BR38", "BR38", SwitchKind.BREAKER, false, false, false, 37, 38);
+
+        createThreeWindingsTransformer(substation, "T3_12", "T3_12", vl1.getId(), vl1.getId(), vl2.getId(),
+                45, 47, 49, 35, 32, 39., 25, 15,
+                5, 7, 9,
+                190, 290, 38,
+                "T3_12", null, ConnectablePosition.Direction.TOP,
+                "T3_12", null, ConnectablePosition.Direction.TOP,
+                "T3_12", null, ConnectablePosition.Direction.TOP);
+
+        return network;
+    }
+
+    public static Network createBusBreakerNetworkWithInternalBranches(String id, String sourceFormat) {
+        Network network = Network.create(id, sourceFormat);
+        Substation substation = createSubstation(network, "S1", "S1", Country.FR);
+
+        VoltageLevel vl1 = createVoltageLevel(substation, "VL1", "VL1", TopologyKind.BUS_BREAKER, 400, -1);
+
+        vl1.getBusBreakerView().newBus()
+                .setId("B11")
+                .add();
+        vl1.newLoad()
+                .setId("LD1")
+                .setConnectableBus("B11")
+                .setBus("B11")
+                .setP0(1.0)
+                .setQ0(1.0)
+                .add();
+
+        network.getVoltageLevel("VL1").getBusBreakerView().newBus()
+                .setId("B12")
+                .add();
+        vl1.getBusBreakerView().newSwitch()
+                .setId("BR1")
+                .setBus1("B12")
+                .setBus2("B11")
+                .setOpen(true)
+                .add();
+
+        vl1.newGenerator()
+                .setId("G")
+                .setBus("B12")
+                .setConnectableBus("B12")
+                .setMaxP(100.0)
+                .setMinP(50.0)
+                .setTargetP(100.0)
+                .setTargetV(400.0)
+                .setVoltageRegulatorOn(true)
+                .add();
+
+        network.newLine()
+                .setId("L11")
+                .setVoltageLevel1("VL1")
+                .setConnectableBus1("B11")
+                .setBus1("B11")
+                .setVoltageLevel2("VL1")
+                .setConnectableBus2("B12")
+                .setBus2("B12")
+                .setR(1.0)
+                .setX(1.0)
+                .setG1(0.0)
+                .setB1(0.0)
+                .setG2(0.0)
+                .setB2(0.0)
+                .add();
+
+        TwoWindingsTransformer twt = substation.newTwoWindingsTransformer()
+                .setId("T11")
+                .setVoltageLevel1("VL1")
+                .setBus1("B11")
+                .setConnectableBus1("B11")
+                .setVoltageLevel2("VL1")
+                .setBus2("B12")
+                .setConnectableBus2("B12")
+                .setR(250)
+                .setX(100)
+                .setG(52)
+                .setB(12)
+                .setRatedU1(65)
+                .setRatedU2(90)
+                .add();
+        twt.newPhaseTapChanger()
+                .setTapPosition(1)
+                .setRegulationTerminal(twt.getTerminal2())
+                .setRegulationMode(PhaseTapChanger.RegulationMode.FIXED_TAP)
+                .setRegulationValue(400)
+                .beginStep()
+                .setAlpha(-20.0)
+                .setRho(1.0)
+                .setR(0.0)
+                .setX(0.0)
+                .setG(0.0)
+                .setB(0.0)
+                .endStep()
+                .beginStep()
+                .setAlpha(0.0)
+                .setRho(1.0)
+                .setR(0.0)
+                .setX(0.0)
+                .setG(0.0)
+                .setB(0.0)
+                .endStep()
+                .beginStep()
+                .setAlpha(20.0)
+                .setRho(1.0)
+                .setR(0.0)
+                .setX(0.0)
+                .setG(0.0)
+                .setB(0.0)
+                .endStep()
+                .add();
+
+        VoltageLevel vl2 = substation.newVoltageLevel()
+                .setId("VL2")
+                .setNominalV(400.0)
+                .setTopologyKind(TopologyKind.BUS_BREAKER)
+                .add();
+        vl2.getBusBreakerView().newBus()
+                .setId("B21")
+                .add();
+        vl2.newLoad()
+                .setId("LD2")
+                .setConnectableBus("B21")
+                .setBus("B21")
+                .setP0(1.0)
+                .setQ0(1.0)
+                .add();
+
+        network.newLine()
+                .setId("L12")
+                .setVoltageLevel1("VL1")
+                .setConnectableBus1("B11")
+                .setBus1("B11")
+                .setVoltageLevel2("VL2")
+                .setConnectableBus2("B21")
+                .setBus2("B21")
+                .setR(1.0)
+                .setX(1.0)
+                .setG1(0.0)
+                .setB1(0.0)
+                .setG2(0.0)
+                .setB2(0.0)
+                .add();
+
+        substation.newTwoWindingsTransformer()
+                .setId("T12")
+                .setVoltageLevel1("VL1")
+                .setBus1("B12")
+                .setConnectableBus1("B12")
+                .setVoltageLevel2("VL2")
+                .setBus2("B21")
+                .setConnectableBus2("B21")
+                .setR(250)
+                .setX(100)
+                .setG(52)
+                .setB(12)
+                .setRatedU1(65)
+                .setRatedU2(90)
+                .add();
+
+        ThreeWindingsTransformer threeWindingsTransformer = substation.newThreeWindingsTransformer()
+                .setId("T3_12")
+                .setRatedU0(234)
+                .newLeg1()
+                .setVoltageLevel("VL1")
+                .setBus("B11")
+                .setConnectableBus("B11")
+                .setR(45)
+                .setX(35)
+                .setG(25)
+                .setB(15)
+                .setRatedU(5)
+                .add()
+                .newLeg2()
+                .setVoltageLevel("VL1")
+                .setBus("B12")
+                .setConnectableBus("B12")
+                .setR(47)
+                .setX(37)
+                .setG(27)
+                .setB(17)
+                .setRatedU(7)
+                .add()
+                .newLeg3()
+                .setVoltageLevel("VL2")
+                .setBus("B21")
+                .setConnectableBus("B21")
+                .setR(49)
+                .setX(39)
+                .setG(29)
+                .setB(19)
+                .setRatedU(9)
+                .add()
+                .add();
+
+        return network;
+    }
+
+    public static Network createNetworkWithSvcVscScDl() {
+        Network network = Network.create("testCase1", "test");
+        Substation substation = network.newSubstation().setId("s").setCountry(Country.FR).add();
+        VoltageLevel vl = substation.newVoltageLevel().setId("vl").setTopologyKind(TopologyKind.NODE_BREAKER).setNominalV(380).add();
+        VoltageLevel.NodeBreakerView view = vl.getNodeBreakerView();
+        BusbarSection bbs = view.newBusbarSection().setId("bbs").setNode(0).add();
+        bbs.newExtension(BusbarSectionPositionAdder.class).withBusbarIndex(1).withSectionIndex(1);
+        BusbarSection bbs2 = view.newBusbarSection().setId("bbs2").setNode(3).add();
+        bbs2.newExtension(BusbarSectionPositionAdder.class).withBusbarIndex(2).withSectionIndex(2);
+        StaticVarCompensator svc = vl.newStaticVarCompensator()
+                .setId("svc")
+                .setName("svc")
+                .setNode(2)
+                .setBmin(0.0002)
+                .setBmax(0.0008)
+                .setRegulationMode(StaticVarCompensator.RegulationMode.VOLTAGE)
+                .setVoltageSetPoint(390)
+                .add();
+        svc.getTerminal()
+                .setP(100.0)
+                .setQ(50.0);
+        VscConverterStation vsc = vl.newVscConverterStation()
+                .setId("vsc")
+                .setName("Converter1")
+                .setNode(1)
+                .setLossFactor(0.011f)
+                .setVoltageSetpoint(405.0)
+                .setVoltageRegulatorOn(true)
+                .add();
+        vsc.getTerminal()
+                .setP(100.0)
+                .setQ(50.0);
+        ShuntCompensator c1 = vl.newShuntCompensator()
+                .setId("C1")
+                .setName("Filter 1")
+                .setNode(4)
+                .setSectionCount(1)
+                .newLinearModel()
+                .setBPerSection(1e-5)
+                .setMaximumSectionCount(1)
+                .add()
+                .add();
+        DanglingLine dl1 = vl.newDanglingLine()
+                .setId("dl1")
+                .setName("Dangling line 1")
+                .setNode(5)
+                .setP0(1)
+                .setQ0(1)
+                .setR(0)
+                .setX(0)
+                .setB(0)
+                .setG(0)
+                .add();
+        dl1.getTerminal()
+                .setP(100.0)
+                .setQ(50.0);
+        view.newDisconnector().setId("d").setNode1(0).setNode2(1).add();
+        view.newBreaker().setId("b").setNode1(1).setNode2(2).add();
+        view.newBreaker().setId("b2").setNode1(3).setNode2(4).add();
+        view.newBreaker().setId("b3").setNode1(3).setNode2(5).add();
+        view.newBreaker().setId("bt").setNode1(0).setNode2(3).add();
+
+        VoltageLevel vl2 = substation.newVoltageLevel().setId("vl2").setTopologyKind(TopologyKind.NODE_BREAKER).setNominalV(380).add();
+        VoltageLevel.NodeBreakerView view2 = vl2.getNodeBreakerView();
+        BusbarSection bbs1Vl2 = view2.newBusbarSection().setId("bbs1Vl2").setNode(0).add();
+        bbs1Vl2.newExtension(BusbarSectionPositionAdder.class).withBusbarIndex(1).withSectionIndex(1);
+        BusbarSection bbs2Vl2 = view2.newBusbarSection().setId("bbs2Vl2").setNode(3).add();
+        bbs2Vl2.newExtension(BusbarSectionPositionAdder.class).withBusbarIndex(2).withSectionIndex(2);
+
+        VscConverterStation vsc2 = vl2.newVscConverterStation()
+                .setId("vsc2")
+                .setName("Converter2")
+                .setNode(1)
+                .setLossFactor(0.011f)
+                .setVoltageSetpoint(405.0)
+                .setVoltageRegulatorOn(true)
+                .add();
+        vsc2.getTerminal()
+                .setP(300.0)
+                .setQ(75.0);
+        view2.newDisconnector().setId("d2").setNode1(0).setNode2(1).add();
+
+        network.newHvdcLine()
+                .setId("hvdc")
+                .setName("hvdc")
+                .setR(1)
+                .setNominalV(380)
+                .setMaxP(200)
+                .setActivePowerSetpoint(150)
+                .setConvertersMode(HvdcLine.ConvertersMode.SIDE_1_INVERTER_SIDE_2_RECTIFIER)
+                .setConverterStationId1("vsc")
+                .setConverterStationId2("vsc2")
+                .add();
+
+        return network;
+    }
+
+    public static Network createNetworkWithFiveBusesFourLoads() {
+        Network network = createNetworkWithTwoParallelLoads();
+        VoltageLevel vl = network.getVoltageLevel("vl1");
+
+        createBusBarSection(vl, "bbs1", "bbs1", 0, 1, 1);
+        createBusBarSection(vl, "bbs21", "bbs21", 1, 2, 1);
+        createBusBarSection(vl, "bbs22", "bbs22", 2, 2, 2);
+        createSwitch(vl, "bA", "bA", SwitchKind.BREAKER, false, false, false, 3, 4);
+        createLoad(vl, "loadA", "loadA", "loadA", null, ConnectablePosition.Direction.TOP, 4, 10, 10);
+        createSwitch(vl, "dA1", "dA1", SwitchKind.DISCONNECTOR, false, false, false, 0, 3);
+        createSwitch(vl, "dA2", "dA2", SwitchKind.DISCONNECTOR, false, false, false, 1, 3);
+
+        createSwitch(vl, "bB", "bB", SwitchKind.BREAKER, false, false, false, 5, 6);
+        createLoad(vl, "loadB", "loadB", "loadB", null, ConnectablePosition.Direction.TOP, 6, 10, 10);
+        createSwitch(vl, "dB1", "dB1", SwitchKind.DISCONNECTOR, false, false, false, 2, 5);
+        createSwitch(vl, "dB2", "dB2", SwitchKind.DISCONNECTOR, false, false, false, 0, 5);
+
+        createSwitch(vl, "link", "link", SwitchKind.BREAKER, false, false, false, 5, 9);
+
+        return network;
+    }
+
+    public static Network createNetworkWithTwoParallelLoads() {
+        Network network = Network.create("TestSingleLineDiagramClass", "test");
+        Substation substation = createSubstation(network, "s", "s", Country.FR);
+        VoltageLevel vl = createVoltageLevel(substation, "vl1", "vl1", TopologyKind.NODE_BREAKER, 380, 10);
+        createBusBarSection(vl, "bbs13", "bbs13", 7, 1, 3);
+        createBusBarSection(vl, "bbs23", "bbs23", 8, 2, 3);
+        createLoad(vl, "loadC", "loadC", "loadC", null, ConnectablePosition.Direction.TOP, 9, 10, 10);
+        createSwitch(vl, "bCD1", "bCD1", SwitchKind.BREAKER, false, false, false, 8, 9);
+        createSwitch(vl, "bCD2", "bCD2", SwitchKind.BREAKER, false, false, false, 7, 9);
+        createSwitch(vl, "bCD3", "bCD3", SwitchKind.BREAKER, false, false, false, 7, 9);
+        createSwitch(vl, "bD1", "bD1", SwitchKind.BREAKER, false, false, false, 20, 9);
+        createLoad(vl, "loadD", "loadD", "loadD", null, ConnectablePosition.Direction.TOP, 20, 10, 10);
+        return network;
+    }
+
+    public static Network createNetworkWithFourParallelLegs() {
+        Network network = Network.create("networkWithFourParallelLegs", "test");
+        Substation substation = createSubstation(network, "s", "s", Country.FR);
+        VoltageLevel vl = createVoltageLevel(substation, "vl1", "vl1", TopologyKind.NODE_BREAKER, 380, 10);
+        createBusBarSection(vl, "bbs1", "bbs1", 0, 1, 3);
+        createBusBarSection(vl, "bbs2", "bbs2", 1, 2, 3);
+        createSwitch(vl, "d1a", "d1an", SwitchKind.DISCONNECTOR, false, false, false, 0, 2);
+        createSwitch(vl, "d2a", "d2an", SwitchKind.DISCONNECTOR, false, false, false, 1, 2);
+        createSwitch(vl, "b1", "b1n", SwitchKind.BREAKER, false, false, false, 2, 3);
+        createSwitch(vl, "d1b", "d1bn", SwitchKind.DISCONNECTOR, false, false, false, 0, 3);
+        createSwitch(vl, "d2b", "d2bn", SwitchKind.DISCONNECTOR, false, false, false, 1, 3);
+        createSwitch(vl, "bl", "bln", SwitchKind.BREAKER, false, false, false, 3, 4);
+        createLoad(vl, "load", "load", "load displayed name", null, ConnectablePosition.Direction.BOTTOM, 4, 10, 10);
+        return network;
+    }
+
+    public static Network createNetworkWithInternalPst() {
+        Network network = Network.create("networkWithInternalPst", "test");
+        Substation substation = createSubstation(network, "s", "s", Country.FR);
+        VoltageLevel vl = createVoltageLevel(substation, "vl1", "vl1", TopologyKind.NODE_BREAKER, 380, 10);
+        createBusBarSection(vl, "bbs1", "bbs1", 0, 1, 3);
+        createBusBarSection(vl, "bbs2", "bbs2", 1, 2, 3);
+        createSwitch(vl, "d1a", "d1an", SwitchKind.DISCONNECTOR, false, false, false, 0, 2);
+        createSwitch(vl, "d2a", "d2an", SwitchKind.DISCONNECTOR, false, false, false, 1, 2);
+        createSwitch(vl, "bpst", "bpstn", SwitchKind.BREAKER, false, false, false, 2, 3);
+        createPhaseShiftTransformer(substation, "trf3", "trf3", 1.0, 14.745, 0.0, 3.2E-5, 380.0, 380.0,
+                3, 4, vl.getId(), vl.getId(),
+                "pst1a", 1, ConnectablePosition.Direction.BOTTOM,
+                "pst1b", 1, ConnectablePosition.Direction.BOTTOM);
+        createSwitch(vl, "dpst", "dpstn", SwitchKind.BREAKER, false, true, false, 4, 5);
+        createSwitch(vl, "d1b", "d1bn", SwitchKind.DISCONNECTOR, false, true, false, 0, 5);
+        createSwitch(vl, "d2b", "d2bn", SwitchKind.DISCONNECTOR, false, true, false, 1, 5);
+        createSwitch(vl, "bl", "bln", SwitchKind.BREAKER, false, false, false, 5, 6);
+        createLoad(vl, "load", "load", "l", null, ConnectablePosition.Direction.BOTTOM, 6, 10, 10);
+        return network;
+    }
+
+    public static Network createNetworkWithFlatSections() {
+        Network network = Network.create("TestSingleLineDiagramClass", "test");
+        Substation substation = createSubstation(network, "s", "s", Country.FR);
+        VoltageLevel vl = createVoltageLevel(substation, "vl1", "vl1", TopologyKind.NODE_BREAKER, 380, 10);
+        createBusBarSection(vl, "bbs11", "bbs11", 0, 1, 1);
+        createBusBarSection(vl, "bbs21", "bbs21", 1, 2, 1);
+        createBusBarSection(vl, "bbs12", "bbs12", 2, 1, 2);
+        createBusBarSection(vl, "bbs22", "bbs22", 3, 2, 2);
+        createBusBarSection(vl, "bbs13", "bbs13", 4, 1, 3);
+        createBusBarSection(vl, "bbs23", "bbs23", 5, 2, 3);
+        createSwitch(vl, "d112", "d112", SwitchKind.DISCONNECTOR, false, false, false, 0, 2);
+        createSwitch(vl, "d212", "d212", SwitchKind.DISCONNECTOR, false, false, false, 1, 3);
+        createSwitch(vl, "d123a", "d123a", SwitchKind.DISCONNECTOR, false, false, false, 2, 6);
+        createSwitch(vl, "b123", "b123", SwitchKind.BREAKER, false, false, false, 6, 7);
+        createSwitch(vl, "d123b", "d123b", SwitchKind.DISCONNECTOR, false, false, false, 7, 4);
+        createSwitch(vl, "d223a", "d223a", SwitchKind.DISCONNECTOR, false, false, false, 3, 8);
+        createSwitch(vl, "b223", "b223", SwitchKind.BREAKER, false, false, false, 8, 9);
+        createSwitch(vl, "d223b", "d223b", SwitchKind.DISCONNECTOR, false, false, false, 9, 5);
+        return network;
+    }
+
+    public static Network createNetworkWithBatteries() {
+        Network network = Network.create("TestBatteries", "test");
+        Substation substation = createSubstation(network, "s", "s", Country.FR);
+        VoltageLevel vl = createVoltageLevel(substation, "vl1", "vl1", TopologyKind.NODE_BREAKER, 380, 10);
+        createBusBarSection(vl, "bbs11", "bbs11", 0, 1, 1);
+        createSwitch(vl, "d1b", "d1b", SwitchKind.DISCONNECTOR, true, false, false, 1, 2);
+        createSwitch(vl, "b1", "b1", SwitchKind.BREAKER, true, false, false, 0, 1);
+        createBattery(vl, "batt1", "batt1", "batt1", null, ConnectablePosition.Direction.TOP, 2, 1, 10, 5, 5);
+        createSwitch(vl, "d2b", "d2b", SwitchKind.DISCONNECTOR, true, true, false, 3, 4);
+        createSwitch(vl, "b2", "b2", SwitchKind.BREAKER, true, true, false, 0, 3);
+        createBattery(vl, "batt2", "batt2", "batt2", null, ConnectablePosition.Direction.BOTTOM, 4, 3, 10, 6, 6);
+        return network;
+    }
+
+    public static Substation createSubstation(Network n, String id, String name, Country country) {
+        return n.newSubstation()
+                .setId(id)
+                .setName(name)
+                .setCountry(country)
+                .add();
+    }
+
+    public static VoltageLevel createVoltageLevel(Substation s, String id, String name,
+                                                  TopologyKind topology, double vNom, int nodeCount) {
+        VoltageLevel vl = s.newVoltageLevel()
+                .setId(id)
+                .setName(name)
+                .setTopologyKind(topology)
+                .setNominalV(vNom)
+                .add();
+        return vl;
+    }
+
+    public static void createSwitch(VoltageLevel vl, String id, String name, SwitchKind kind, boolean retained, boolean open, boolean fictitious, int node1, int node2) {
+        vl.getNodeBreakerView().newSwitch()
+                .setId(id)
+                .setName(name)
+                .setKind(kind)
+                .setRetained(retained)
+                .setOpen(open)
+                .setFictitious(fictitious)
+                .setNode1(node1)
+                .setNode2(node2)
+                .add();
+    }
+
+    public static void createInternalConnection(VoltageLevel vl, int node1, int node2) {
+        vl.getNodeBreakerView().newInternalConnection()
+                .setNode1(node1)
+                .setNode2(node2)
+                .add();
+    }
+
+    public static void createBusBarSection(VoltageLevel vl, String id, String name, int node, int busbarIndex, int sectionIndex) {
+        BusbarSection bbs = vl.getNodeBreakerView().newBusbarSection()
+                .setId(id)
+                .setName(name)
+                .setNode(node)
+                .add();
+        bbs.newExtension(BusbarSectionPositionAdder.class)
+                .withBusbarIndex(busbarIndex)
+                .withSectionIndex(sectionIndex)
+                .add();
+    }
+
+    public static void createLoad(VoltageLevel vl, String id, String name, String feederName, Integer feederOrder,
+                                  ConnectablePosition.Direction direction, int node, double p0, double q0) {
+        Load load = vl.newLoad()
+                .setId(id)
+                .setName(name)
+                .setNode(node)
+                .setP0(p0)
+                .setQ0(q0)
+                .add();
+        addFeederPosition(load, feederName, feederOrder, direction);
+    }
+
+    public static void createGenerator(VoltageLevel vl, String id, String name, String feederName, Integer feederOrder,
+                                       ConnectablePosition.Direction direction, int node,
+                                       double minP, double maxP, boolean voltageRegulator,
+                                       double targetP, double targetQ) {
+        Generator gen = vl.newGenerator()
+                .setId(id)
+                .setName(name)
+                .setNode(node)
+                .setMinP(minP)
+                .setMaxP(maxP)
+                .setVoltageRegulatorOn(voltageRegulator)
+                .setTargetP(targetP)
+                .setTargetQ(targetQ)
+                .add();
+        addFeederPosition(gen, feederName, feederOrder, direction);
+    }
+
+    protected static void createBattery(VoltageLevel vl, String id, String name, String feederName, Integer feederOrder,
+                                        ConnectablePosition.Direction direction, int node,
+                                        double minP, double maxP,
+                                        double targetP, double targetQ) {
+        Battery battery = vl.newBattery()
+                .setId(id)
+                .setName(name)
+                .setNode(node)
+                .setMinP(minP)
+                .setMaxP(maxP)
+                .setTargetP(targetP)
+                .setTargetQ(targetQ)
+                .add();
+        addFeederPosition(battery, feederName, feederOrder, direction);
+    }
+
+    public static void createShunt(VoltageLevel vl, String id, String name, String feederName, Integer feederOrder,
+                                   ConnectablePosition.Direction direction, int node,
+                                   double bPerSection, int maximumSectionCount, int currentSectionCount) {
+        ShuntCompensator shunt = vl.newShuntCompensator()
+                .setId(id)
+                .setName(name)
+                .setNode(node)
+                .setSectionCount(currentSectionCount)
+                .newLinearModel()
+                .setBPerSection(bPerSection)
+                .setMaximumSectionCount(maximumSectionCount)
+                .add()
+                .add();
+        addFeederPosition(shunt, feederName, feederOrder, direction);
+    }
+
+    public static TwoWindingsTransformer createTwoWindingsTransformer(Substation s, String id, String name,
+                                                                      double r, double x, double g, double b,
+                                                                      double ratedU1, double ratedU2,
+                                                                      int node1, int node2,
+                                                                      String idVoltageLevel1, String idVoltageLevel2,
+                                                                      String feederName1, Integer feederOrder1, ConnectablePosition.Direction direction1,
+                                                                      String feederName2, Integer feederOrder2, ConnectablePosition.Direction direction2) {
+        TwoWindingsTransformer t = s.newTwoWindingsTransformer()
+                .setId(id)
+                .setName(name)
+                .setR(r)
+                .setX(x)
+                .setG(g)
+                .setB(b)
+                .setRatedU1(ratedU1)
+                .setRatedU2(ratedU2)
+                .setNode1(node1)
+                .setVoltageLevel1(idVoltageLevel1)
+                .setNode2(node2)
+                .setVoltageLevel2(idVoltageLevel2)
+                .add();
+        addTwoFeedersPosition(t, feederName1, feederOrder1, direction1, feederName2, feederOrder2, direction2);
+        return t;
+    }
+
+    protected static void createPhaseShiftTransformer(Substation s, String id, String name,
+                                                      double r, double x, double g, double b,
+                                                      double ratedU1, double ratedU2,
+                                                      int node1, int node2,
+                                                      String idVoltageLevel1, String idVoltageLevel2,
+                                                      String feederName1, Integer feederOrder1, ConnectablePosition.Direction direction1,
+                                                      String feederName2, Integer feederOrder2, ConnectablePosition.Direction direction2) {
+        TwoWindingsTransformer twt = createTwoWindingsTransformer(s, id, name, r, x, g, b, ratedU1, ratedU2, node1, node2,
+                idVoltageLevel1, idVoltageLevel2, feederName1, feederOrder1, direction1, feederName2, feederOrder2, direction2);
+        twt.newPhaseTapChanger()
+                .setTapPosition(1)
+                .setRegulationTerminal(twt.getTerminal2())
+                .setRegulationMode(PhaseTapChanger.RegulationMode.FIXED_TAP)
+                .setRegulationValue(200)
+                .beginStep()
+                .setAlpha(-20.0)
+                .setRho(1.0)
+                .setR(0.0)
+                .setX(0.0)
+                .setG(0.0)
+                .setB(0.0)
+                .endStep()
+                .beginStep()
+                .setAlpha(0.0)
+                .setRho(1.0)
+                .setR(0.0)
+                .setX(0.0)
+                .setG(0.0)
+                .setB(0.0)
+                .endStep()
+                .beginStep()
+                .setAlpha(20.0)
+                .setRho(1.0)
+                .setR(0.0)
+                .setX(0.0)
+                .setG(0.0)
+                .setB(0.0)
+                .endStep()
+                .add();
+    }
+
+    public static void createThreeWindingsTransformer(Substation s, String id, String name,
+                                                      String vl1, String vl2, String vl3,
+                                                      double r1, double r2, double r3,
+                                                      double x1, double x2, double x3,
+                                                      double g1, double b1,
+                                                      double ratedU1, double ratedU2, double ratedU3,
+                                                      int node1, int node2, int node3,
+                                                      String feederName1, Integer feederOrder1, ConnectablePosition.Direction direction1,
+                                                      String feederName2, Integer feederOrder2, ConnectablePosition.Direction direction2,
+                                                      String feederName3, Integer feederOrder3, ConnectablePosition.Direction direction3) {
+        ThreeWindingsTransformer t = s.newThreeWindingsTransformer()
+                .setId(id)
+                .setName(name)
+                .newLeg1()
+                .setR(r1)
+                .setX(x1)
+                .setG(g1)
+                .setB(b1)
+                .setRatedU(ratedU1)
+                .setVoltageLevel(vl1)
+                .setNode(node1)
+                .add()
+                .newLeg2()
+                .setR(r2)
+                .setX(x2)
+                .setRatedU(ratedU2)
+                .setVoltageLevel(vl2)
+                .setNode(node2)
+                .add()
+                .newLeg3()
+                .setR(r3)
+                .setX(x3)
+                .setRatedU(ratedU3)
+                .setVoltageLevel(vl3)
+                .setNode(node3)
+                .add()
+                .add();
+
+        addThreeFeedersPosition(t, feederName1, feederOrder1, direction1, feederName2, feederOrder2, direction2, feederName3, feederOrder3, direction3);
+    }
+
+    public static void createLine(Network network,
+                                  String id, String name,
+                                  double r, double x,
+                                  double g1, double b1,
+                                  double g2, double b2,
+                                  int node1, int node2,
+                                  String idVoltageLevel1, String idVoltageLevel2,
+                                  String feederName1, Integer feederOrder1, ConnectablePosition.Direction direction1,
+                                  String feederName2, Integer feederOrder2, ConnectablePosition.Direction direction2) {
+        Line line = network.newLine()
+                .setId(id)
+                .setName(name)
+                .setR(r)
+                .setX(x)
+                .setG1(g1)
+                .setB1(b1)
+                .setG2(g2)
+                .setB2(b2)
+                .setNode1(node1)
+                .setVoltageLevel1(idVoltageLevel1)
+                .setNode2(node2)
+                .setVoltageLevel2(idVoltageLevel2)
+                .add();
+        addTwoFeedersPosition(line, feederName1, feederOrder1, direction1, feederName2, feederOrder2, direction2);
+    }
+
+    private static void addFeederPosition(Extendable<?> extendable, String feederName, Integer feederOrder, ConnectablePosition.Direction direction) {
+        ConnectablePositionAdder.FeederAdder feederAdder = extendable.newExtension(ConnectablePositionAdder.class).newFeeder();
+        if (feederOrder != null) {
+            feederAdder.withOrder(feederOrder);
+        }
+        feederAdder.withDirection(direction).withName(feederName).add()
+                .add();
+    }
+
+    private static void addTwoFeedersPosition(Extendable<?> extendable,
+                                              String feederName1, Integer feederOrder1, ConnectablePosition.Direction direction1,
+                                              String feederName2, Integer feederOrder2, ConnectablePosition.Direction direction2) {
+        ConnectablePositionAdder extensionAdder = extendable.newExtension(ConnectablePositionAdder.class);
+        ConnectablePositionAdder.FeederAdder feederAdder1 = extensionAdder.newFeeder1();
+        if (feederOrder1 != null) {
+            feederAdder1.withOrder(feederOrder1);
+        }
+        feederAdder1.withName(feederName1).withDirection(direction1).add();
+        ConnectablePositionAdder.FeederAdder feederAdder2 = extensionAdder.newFeeder2();
+        if (feederOrder2 != null) {
+            feederAdder2.withOrder(feederOrder2);
+        }
+        feederAdder2.withName(feederName2).withDirection(direction2).add();
+        extensionAdder.add();
+    }
+
+    private static void addThreeFeedersPosition(Extendable<?> extendable,
+                                                String feederName1, Integer feederOrder1, ConnectablePosition.Direction direction1,
+                                                String feederName2, Integer feederOrder2, ConnectablePosition.Direction direction2,
+                                                String feederName3, Integer feederOrder3, ConnectablePosition.Direction direction3) {
+        ConnectablePositionAdder extensionAdder = extendable.newExtension(ConnectablePositionAdder.class);
+        ConnectablePositionAdder.FeederAdder feederAdder1 = extensionAdder.newFeeder1();
+        if (feederOrder1 != null) {
+            feederAdder1.withOrder(feederOrder1);
+        }
+        feederAdder1.withName(feederName1).withDirection(direction1).add();
+        ConnectablePositionAdder.FeederAdder feederAdder2 = extensionAdder.newFeeder2();
+        if (feederOrder2 != null) {
+            feederAdder2.withOrder(feederOrder2);
+        }
+        feederAdder2.withName(feederName2).withDirection(direction2).add();
+        ConnectablePositionAdder.FeederAdder feederAdder3 = extensionAdder.newFeeder3();
+        if (feederOrder3 != null) {
+            feederAdder3.withOrder(feederOrder3);
+        }
+        feederAdder3.withName(feederName3).withDirection(direction3).add();
+        extensionAdder.add();
     }
 }
