+++ single-line-diagram/single-line-diagram-core/src/test/java/com/powsybl/sld/iidm/TestBattery.java
@@ -6,6 +6,7 @@
  */
 package com.powsybl.sld.iidm;
 
+import com.powsybl.diagram.test.Networks;
 import com.powsybl.sld.builders.NetworkGraphBuilder;
 import com.powsybl.sld.model.graphs.VoltageLevelGraph;
 import org.junit.jupiter.api.BeforeEach;
@@ -22,7 +23,7 @@ class TestBattery extends AbstractTestCaseIidm {
 
     @BeforeEach
     public void setUp() throws IOException {
-        network = CreateNetworksUtil.createNetworkWithBatteries();
+        network = Networks.createNetworkWithBatteries();
         graphBuilder = new NetworkGraphBuilder(network);
     }
