+++ single-line-diagram/single-line-diagram-core/src/test/java/com/powsybl/sld/iidm/TestCase10TestBreakerToBus.java
@@ -6,6 +6,7 @@
  */
 package com.powsybl.sld.iidm;
 
+import com.powsybl.diagram.test.Networks;
 import com.powsybl.iidm.network.Country;
 import com.powsybl.iidm.network.Network;
 import com.powsybl.iidm.network.SwitchKind;
@@ -33,10 +34,10 @@ class TestCase10TestBreakerToBus extends AbstractTestCaseIidm {
     @Override
     public void setUp() {
         network = Network.create("testCase1", "AbstractTest");
-        substation = createSubstation(network, "s", "s", Country.FR);
-        vl = createVoltageLevel(substation, "vl", "vl", TopologyKind.NODE_BREAKER, 380, 10);
-        createBusBarSection(vl, "bbs", "bbs", 0, 1, 1);
-        createLoad(vl, "l", "l", "l", 0, ConnectablePosition.Direction.TOP, 2, 10, 10);
-        createSwitch(vl, "b", "b", SwitchKind.BREAKER, false, false, false, 0, 2);
+        substation = Networks.createSubstation(network, "s", "s", Country.FR);
+        vl = Networks.createVoltageLevel(substation, "vl", "vl", TopologyKind.NODE_BREAKER, 380, 10);
+        Networks.createBusBarSection(vl, "bbs", "bbs", 0, 1, 1);
+        Networks.createLoad(vl, "l", "l", "l", 0, ConnectablePosition.Direction.TOP, 2, 10, 10);
+        Networks.createSwitch(vl, "b", "b", SwitchKind.BREAKER, false, false, false, 0, 2);
     }
 }
