+++ single-line-diagram/single-line-diagram-core/src/test/java/com/powsybl/sld/iidm/TestCase11SubstationGraph.java
@@ -6,6 +6,7 @@
  */
 package com.powsybl.sld.iidm;
 
+import com.powsybl.diagram.test.Networks;
 import com.powsybl.iidm.network.extensions.ConnectablePosition;
 import com.powsybl.sld.builders.NetworkGraphBuilder;
 import com.powsybl.sld.layout.HorizontalSubstationLayoutFactory;
@@ -30,7 +31,7 @@ class TestCase11SubstationGraph extends AbstractTestCaseIidm {
 
     @BeforeEach
     public void setUp() {
-        network = NetworkFactory.createTestCase11Network();
+        network = Networks.createTestCase11Network();
         substation = network.getSubstation("subst");
         graphBuilder = new NetworkGraphBuilder(network);
     }
@@ -95,7 +96,7 @@ class TestCase11SubstationGraph extends AbstractTestCaseIidm {
     void testRight3wtOrientation() {
         // build substation graph
         network.getThreeWindingsTransformer("trf7").remove();
-        createThreeWindingsTransformer(substation, "trf7", "trf7", "vl3", "vl2", "vl1",
+        Networks.createThreeWindingsTransformer(substation, "trf7", "trf7", "vl3", "vl2", "vl1",
                 0.5, 0.5, 0.5, 1., 1., 1., 0.1, 0.1,
                 50., 225., 400.,
                 8, 19, 31,
@@ -115,7 +116,7 @@ class TestCase11SubstationGraph extends AbstractTestCaseIidm {
     void testLeft3wtOrientation() {
         // build substation graph
         network.getThreeWindingsTransformer("trf7").remove();
-        createThreeWindingsTransformer(substation, "trf7", "trf7", "vl2", "vl1", "vl3",
+        Networks.createThreeWindingsTransformer(substation, "trf7", "trf7", "vl2", "vl1", "vl3",
                 0.5, 0.5, 0.5, 1., 1., 1., 0.1, 0.1,
                 225., 400., 50.,
                 19, 31, 8,
