+++ single-line-diagram/single-line-diagram-core/src/test/java/com/powsybl/sld/iidm/TestCase5ShuntVertical.java
@@ -6,6 +6,7 @@
  */
 package com.powsybl.sld.iidm;
 
+import com.powsybl.diagram.test.Networks;
 import com.powsybl.iidm.network.Country;
 import com.powsybl.iidm.network.Network;
 import com.powsybl.iidm.network.SwitchKind;
@@ -42,16 +43,16 @@ class TestCase5ShuntVertical extends AbstractTestCaseIidm {
     public void setUp() {
         network = Network.create("testCase1", "test");
         graphBuilder = new NetworkGraphBuilder(network);
-        substation = createSubstation(network, "s", "s", Country.FR);
-        vl = createVoltageLevel(substation, "vl", "vl", TopologyKind.NODE_BREAKER, 380, 10);
-        createBusBarSection(vl, "bbs", "bbs", 0, 1, 1);
-        createLoad(vl, "la", "la", "la", 20, ConnectablePosition.Direction.TOP, 2, 10, 10);
-        createSwitch(vl, "ba", "ba", SwitchKind.BREAKER, false, false, false, 2, 1);
-        createSwitch(vl, "da", "da", SwitchKind.DISCONNECTOR, false, false, false, 1, 0);
-        createLoad(vl, "lb", "lb", "lb", 10, ConnectablePosition.Direction.TOP, 4, 10, 10);
-        createSwitch(vl, "bb", "bb", SwitchKind.BREAKER, false, false, false, 4, 3);
-        createSwitch(vl, "db", "db", SwitchKind.DISCONNECTOR, false, false, false, 3, 0);
-        createSwitch(vl, "bs", "bs", SwitchKind.BREAKER, false, false, false, 2, 3);
+        substation = Networks.createSubstation(network, "s", "s", Country.FR);
+        vl = Networks.createVoltageLevel(substation, "vl", "vl", TopologyKind.NODE_BREAKER, 380, 10);
+        Networks.createBusBarSection(vl, "bbs", "bbs", 0, 1, 1);
+        Networks.createLoad(vl, "la", "la", "la", 20, ConnectablePosition.Direction.TOP, 2, 10, 10);
+        Networks.createSwitch(vl, "ba", "ba", SwitchKind.BREAKER, false, false, false, 2, 1);
+        Networks.createSwitch(vl, "da", "da", SwitchKind.DISCONNECTOR, false, false, false, 1, 0);
+        Networks.createLoad(vl, "lb", "lb", "lb", 10, ConnectablePosition.Direction.TOP, 4, 10, 10);
+        Networks.createSwitch(vl, "bb", "bb", SwitchKind.BREAKER, false, false, false, 4, 3);
+        Networks.createSwitch(vl, "db", "db", SwitchKind.DISCONNECTOR, false, false, false, 3, 0);
+        Networks.createSwitch(vl, "bs", "bs", SwitchKind.BREAKER, false, false, false, 2, 3);
     }
 
     @Test
