+++ single-line-diagram/single-line-diagram-core/src/test/java/com/powsybl/sld/iidm/TestDoubleForkNode.java
@@ -6,6 +6,7 @@
  */
 package com.powsybl.sld.iidm;
 
+import com.powsybl.diagram.test.Networks;
 import com.powsybl.sld.builders.NetworkGraphBuilder;
 import com.powsybl.sld.model.graphs.VoltageLevelGraph;
 
@@ -23,7 +24,7 @@ class TestDoubleForkNode extends AbstractTestCaseIidm {
 
     @BeforeEach
     public void setUp() throws IOException {
-        network = CreateNetworksUtil.createNetworkWithTwoParallelLoads();
+        network = Networks.createNetworkWithTwoParallelLoads();
         graphBuilder = new NetworkGraphBuilder(network);
     }
